+++ src/main/java/factory/AccountFactory.java
@@ -1,3 +1,7 @@
+package factory;
+
+import factory.Account;
+
 public interface AccountFactory {
     Account createAccount(String type);
 }
