+++ src/main/java/factory/CheckingAccount.java
@@ -1,8 +1,12 @@
-public class SavingsAccount implements Account {
+package factory;
+
+import factory.Account;
+
+public class CheckingAccount implements Account {
     private String accountId;
     private double balance;
 
-    public SavingsAccount(String accountId) {
+    public CheckingAccount(String accountId) {
         this.accountId = accountId;
         this.balance = 0.0;
     }
@@ -14,7 +18,7 @@ public class SavingsAccount implements Account {
 
     @Override
     public String getAccountType() {
-        return "SAVINGS";
+        return "CHECKING";
     }
 
     @Override
