+++ src/main/java/factory/CheckingAccountFactory.java
@@ -1,3 +1,5 @@
+package factory;
+
 public class CheckingAccountFactory implements AccountFactory {
     @Override
     public Account createAccount(String accountId) {
