+++ src/main/java/observer/Observer.java
@@ -1,3 +1,5 @@
+package observer;
+
 public interface Observer {
     void update(String message);
 }
