+++ src/main/java/observer/Subject.java
@@ -1,3 +1,7 @@
+package observer;
+
+import observer.Observer;
+
 public interface Subject {
     void attach(Observer o);
     void detach(Observer o);
