+++ src/main/java/observer/Transaction.java
@@ -1,3 +1,5 @@
+package observer;
+
 import java.sql.Timestamp;
 
 public class Transaction {
@@ -19,7 +21,7 @@ public class Transaction {
 
     // Other methods if necessary
     public void print() {
-        System.out.println("Transaction ID: " + transactionId + ", Account ID: " + accountId + ", Type: " + type + ", Amount: " + amount + ", Timestamp: " + timestamp);
+        System.out.println("observer.Transaction ID: " + transactionId + ", Factory.Account ID: " + accountId + ", Type: " + type + ", Amount: " + amount + ", Timestamp: " + timestamp);
     }
 
 }
